Moved final extensions to new Ext Lib
